Fix: change all integer field to decimal field

+++ src/foods/models.py
@@ -16,8 +16,8 @@ class Food(BaseModel):
     title = models.CharField(max_length=50, unique=True)
     description = models.TextField(blank=True, null=True)
     image = models.ImageField(upload_to = 'images/foods/', default="/images/foods/default.jpg")
-    price = models.FloatField()
-    discount = models.FloatField(default=0.0)
+    price = models.DecimalField()
+    discount = models.DecimalField(default=0.0)
     category = models.ForeignKey(Category, on_delete=models.PROTECT)
 
     def __str__(self) -> str:
